feat(condo): DOMA-7446 move `invoice rows` field into external module
