+++ apps/condo/domains/marketplace/schema/Invoice.js
@@ -2,24 +2,20 @@
  * Generated by `createschema marketplace.Invoice 'number:Integer; property:Relationship:Property:PROTECT; unitType:Text; unitName:Text; accountNumber:Text; toPay:Decimal; items:Json; contact?:Relationship:Contact:SET_NULL; client?:Relationship:User:SET_NULL; clientName?:Text; clientPhone?:Text; clientEmail?:Text'`
  */
 
-const Ajv = require('ajv')
-const addFormats = require('ajv-formats')
-
 const { GQLError, GQLErrorCode: { BAD_USER_INPUT } } = require('@open-condo/keystone/errors')
 const { AutoIncrementInteger } = require('@open-condo/keystone/fields')
 const { historical, versioned, uuided, tracked, softDeleted, dvAndSender } = require('@open-condo/keystone/plugins')
 const { GQLListSchema } = require('@open-condo/keystone/schema')
 
 const { MONEY_AMOUNT_FIELD } = require('@condo/domains/common/schema/fields')
-const { render, getGQLErrorValidator } = require('@condo/domains/common/schema/json.utils')
 const access = require('@condo/domains/marketplace/access/Invoice')
 const {
-    ERROR_INVALID_INVOICE_ROWS,
     ERROR_NO_INVOICE_RECEIVERS,
     ERROR_INVOICE_ALREADY_PAID,
     INVOICE_STATUSES,
     INVOICE_STATUS_DRAFT, INVOICE_STATUS_PAID,
 } = require('@condo/domains/marketplace/constants')
+const { INVOICE_ROWS_FIELD } = require('@condo/domains/marketplace/schema/fields/invoiceRows')
 
 const ERRORS = {
     NO_RECEIVERS: {
@@ -35,50 +31,6 @@ const ERRORS = {
         messageForUser: 'api.marketplace.invoice.error.alreadyPaid',
     },
 }
-
-const INVOICE_ROW_GQL_TYPE_NAME = 'InvoiceRowSchemaField'
-const INVOICE_ROW_GQL_INPUT_NAME = 'InvoiceRowSchemaFieldInput'
-
-const invoiceRowSchemaFields = {
-    name: 'String!',
-    toPay: 'String!',
-    count: 'Int!',
-    vat: 'String',
-    salesTax: 'String',
-    sku: 'String!',
-}
-
-const rowsGqlSchemaTypes = `
-    type ${INVOICE_ROW_GQL_TYPE_NAME} {
-        ${render(invoiceRowSchemaFields)}
-    }
-    
-    input ${INVOICE_ROW_GQL_INPUT_NAME} {
-        ${render(invoiceRowSchemaFields)}
-    }
-`
-
-const ajv = new Ajv()
-addFormats(ajv)
-
-const rowsFieldSchema = {
-    type: 'array',
-    items: {
-        type: 'object',
-        properties: {
-            name: { type: 'string' },
-            toPay: { type: 'string' },
-            count: { type: 'integer' },
-            vat: { type: 'string' },
-            salesTax: { type: 'string' },
-            sku: { type: 'string' },
-        },
-        required: ['name', 'toPay', 'count', 'sku'],
-        additionalProperties: false,
-    },
-}
-
-const validateRowsField = getGQLErrorValidator(ajv.compile(rowsFieldSchema), ERROR_INVALID_INVOICE_ROWS)
 
 const Invoice = new GQLListSchema('Invoice', {
     schemaDoc: 'Invoice model contains information about paid items and payer',
@@ -125,18 +77,7 @@ const Invoice = new GQLListSchema('Invoice', {
 
         toPay: MONEY_AMOUNT_FIELD,
 
-        rows: {
-            schemaDoc: 'The list of paid items',
-            type: 'Json',
-            isRequired: true,
-            hooks: {
-                validateInput: validateRowsField,
-            },
-            extendGraphQLTypes: [rowsGqlSchemaTypes],
-            graphQLInputType: `[${INVOICE_ROW_GQL_INPUT_NAME}!]!`,
-            graphQLReturnType: `[${INVOICE_ROW_GQL_TYPE_NAME}!]!`,
-            graphQLAdminFragment: `{ ${Object.keys(invoiceRowSchemaFields).join(' ')} }`,
-        },
+        rows: INVOICE_ROWS_FIELD,
 
         ticket: {
             schemaDoc: 'The ticket related to this invoice',
